+++ app/vmstorage/main.go
@@ -557,9 +557,6 @@ func registerStorageMetrics() {
 		return float64(m().SearchDelays)
 	})
 
-	metrics.NewGauge(`vm_sorted_row_labels_inserts_total`, func() float64 {
-		return float64(m().SortedRowLabelsInserts)
-	})
 	metrics.NewGauge(`vm_slow_row_inserts_total`, func() float64 {
 		return float64(m().SlowRowInserts)
 	})