+++ lib/storage/storage.go
@@ -48,7 +48,6 @@ type Storage struct {
 	searchTSIDsConcurrencyLimitReached uint64
 	searchTSIDsConcurrencyLimitTimeout uint64
 
-	sortedRowLabelsInserts uint64
 	slowRowInserts         uint64
 	slowPerDayIndexInserts uint64
 	slowMetricNameLoads    uint64
@@ -359,7 +358,6 @@ type Metrics struct {
 
 	SearchDelays uint64
 
-	SortedRowLabelsInserts uint64
 	SlowRowInserts         uint64
 	SlowPerDayIndexInserts uint64
 	SlowMetricNameLoads    uint64
@@ -429,7 +427,6 @@ func (s *Storage) UpdateMetrics(m *Metrics) {
 
 	m.SearchDelays = storagepacelimiter.Search.DelaysTotal()
 
-	m.SortedRowLabelsInserts += atomic.LoadUint64(&s.sortedRowLabelsInserts)
 	m.SlowRowInserts += atomic.LoadUint64(&s.slowRowInserts)
 	m.SlowPerDayIndexInserts += atomic.LoadUint64(&s.slowPerDayIndexInserts)
 	m.SlowMetricNameLoads += atomic.LoadUint64(&s.slowMetricNameLoads)
@@ -1321,8 +1318,6 @@ func (s *Storage) ForceMergePartitions(partitionNamePrefix string) error {
 var rowsAddedTotal uint64
 
 // AddRows adds the given mrs to s.
-//
-// AddRows can modify mrs contents.
 func (s *Storage) AddRows(mrs []MetricRow, precisionBits uint8) error {
 	if len(mrs) == 0 {
 		return nil
@@ -1447,9 +1442,6 @@ func (s *Storage) add(rows []rawRow, mrs []MetricRow, precisionBits uint8) ([]ra
 		prevMetricNameRaw []byte
 	)
 	var pmrs *pendingMetricRows
-	var mn MetricName
-	var metricNameRawSorted []byte
-	var sortedRowLabelsInserts uint64
 	minTimestamp, maxTimestamp := s.tb.getMinMaxTimestamps()
 	// Return only the first error, since it has no sense in returning all errors.
 	var firstWarn error
@@ -1502,40 +1494,22 @@ func (s *Storage) add(rows []rawRow, mrs []MetricRow, precisionBits uint8) ([]ra
 			continue
 		}
 
-		// Slower path - sort labels in MetricNameRaw and check the cache again.
-		// This should limit the number of cache entries for metrics with distinct order of labels to 1.
-		if err := mn.unmarshalRaw(mr.MetricNameRaw); err != nil {
-			if firstWarn == nil {
-				firstWarn = fmt.Errorf("cannot unmarshal MetricNameRaw %q: %w", mr.MetricNameRaw, err)
-			}
-			j--
-			continue
-		}
-		mn.sortTags()
-		metricNameRawSorted = mn.marshalRaw(metricNameRawSorted[:0])
-		if s.getTSIDFromCache(&r.TSID, metricNameRawSorted) {
-			// The TSID for the given metricNameRawSorted has been found in cache and isn't deleted.
-			// There is no need in checking whether r.TSID.MetricID is deleted, since tsidCache doesn't
-			// contain MetricName->TSID entries for deleted time series.
-			// See Storage.DeleteMetrics code for details.
-			sortedRowLabelsInserts++
-			prevTSID = r.TSID
-			prevMetricNameRaw = mr.MetricNameRaw
-			continue
-		}
-
 		// Slow path - the TSID is missing in the cache.
 		// Postpone its search in the loop below.
 		j--
 		if pmrs == nil {
 			pmrs = getPendingMetricRows()
 		}
-		if string(mr.MetricNameRaw) != string(metricNameRawSorted) {
-			mr.MetricNameRaw = append(mr.MetricNameRaw[:0], metricNameRawSorted...)
+		if err := pmrs.addRow(mr); err != nil {
+			// Do not stop adding rows on error - just skip invalid row.
+			// This guarantees that invalid rows don't prevent
+			// from adding valid rows into the storage.
+			if firstWarn == nil {
+				firstWarn = err
+			}
+			continue
 		}
-		pmrs.addRow(mr, &mn)
 	}
-	atomic.AddUint64(&s.sortedRowLabelsInserts, sortedRowLabelsInserts)
 	if pmrs != nil {
 		// Sort pendingMetricRows by canonical metric name in order to speed up search via `is` in the loop below.
 		pendingMetricRows := pmrs.pmrs
@@ -1615,6 +1589,7 @@ type pendingMetricRows struct {
 
 	lastMetricNameRaw []byte
 	lastMetricName    []byte
+	mn                MetricName
 }
 
 func (pmrs *pendingMetricRows) reset() {
@@ -1626,14 +1601,19 @@ func (pmrs *pendingMetricRows) reset() {
 	pmrs.metricNamesBuf = pmrs.metricNamesBuf[:0]
 	pmrs.lastMetricNameRaw = nil
 	pmrs.lastMetricName = nil
+	pmrs.mn.Reset()
 }
 
-func (pmrs *pendingMetricRows) addRow(mr *MetricRow, mn *MetricName) {
+func (pmrs *pendingMetricRows) addRow(mr *MetricRow) error {
 	// Do not spend CPU time on re-calculating canonical metricName during bulk import
 	// of many rows for the same metric.
 	if string(mr.MetricNameRaw) != string(pmrs.lastMetricNameRaw) {
+		if err := pmrs.mn.unmarshalRaw(mr.MetricNameRaw); err != nil {
+			return fmt.Errorf("cannot unmarshal MetricNameRaw %q: %w", mr.MetricNameRaw, err)
+		}
+		pmrs.mn.sortTags()
 		metricNamesBufLen := len(pmrs.metricNamesBuf)
-		pmrs.metricNamesBuf = mn.Marshal(pmrs.metricNamesBuf)
+		pmrs.metricNamesBuf = pmrs.mn.Marshal(pmrs.metricNamesBuf)
 		pmrs.lastMetricName = pmrs.metricNamesBuf[metricNamesBufLen:]
 		pmrs.lastMetricNameRaw = mr.MetricNameRaw
 	}
@@ -1641,6 +1621,7 @@ func (pmrs *pendingMetricRows) addRow(mr *MetricRow, mn *MetricName) {
 		MetricName: pmrs.lastMetricName,
 		mr:         *mr,
 	})
+	return nil
 }
 
 func getPendingMetricRows() *pendingMetricRows {